Training V2 Done
Finally achieved desirable training. Found a public dataset with proper vehicle and number plate images. Currently model is for cars only. Precision and Recall at 95%+, seems to be working on .pt, conversion to .onnx done. Next step is to create the pipeline for onnx and think about deployment in fastAPI or Flask.

Due to problems in LFS, have removed weights from repo and created the helper file for it. Will be using personal Gdrive for any large files going forward.