Onnx Testing
Tested the trained model of Yolov7 after Onnx conversion on  images and seems to be working. Wrote helper functions for pre and post processing. Also got 2 videos to test the data on. Now next i plan to use FastAPI to create a simple endpoint along with zmq to analyze on this data and save the results in file storage and csv (metadata).